Organizing the metastore's unit tests.
